gbaque: GetScouterInfo signed scouterValue load (86.7%)

+++ src/gbaque.cpp
@@ -4403,11 +4403,11 @@ int GbaQueue::GetScouterInfo(int channel, unsigned char* outData)
 					scouterEntry[6] = 0xFF;
 					scouterEntry[7] = 0xFF;
 				} else {
-					const unsigned short scouterValue = *reinterpret_cast<unsigned short*>(enemyEntry + 0xC);
-					if ((static_cast<short>(scouterValue) < 1) || ((scouterValue & 0xC000) == 0x4000)) {
+					const short scouterValue = *reinterpret_cast<short*>(enemyEntry + 0xC);
+					if ((scouterValue < 1) || ((scouterValue & 0xC000) == 0x4000)) {
 						*reinterpret_cast<unsigned short*>(scouterEntry + 6) = 0;
 					} else {
-						*reinterpret_cast<unsigned short*>(scouterEntry + 6) = SwapU16(scouterValue);
+						*reinterpret_cast<unsigned short*>(scouterEntry + 6) = SwapU16(static_cast<unsigned short>(scouterValue));
 					}
 				}
